// apps/web/src/shared/api/gen/model/getPayment200Data.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Checkout API
 * POST /api/sessions с телом {} создаёт гостевую сессию. Вставьте data.token в Authorize без слова Bearer. Суммы указаны в копейках. Данные сохраняются после перезапуска. Оплата тестовая.
 * OpenAPI spec version: 1.0.0
 */
import type { GetPayment200DataCurrency } from './getPayment200DataCurrency';
import type { GetPayment200DataFailureCode } from './getPayment200DataFailureCode';
import type { GetPayment200DataStatus } from './getPayment200DataStatus';

export type GetPayment200Data = {
  id: string;
  orderId: string;
  status: (typeof GetPayment200DataStatus)[keyof typeof GetPayment200DataStatus];
  /**
   * Копейки; 129000 = 1 290 ₽.
   * @minimum 0
   */
  amount: number;
  currency: GetPayment200DataCurrency;
  createdAt: string;
  /** @nullable */
  failureCode: GetPayment200DataFailureCode;
};