// apps/web/src/shared/api/gen/model/getSandbox200DataCardsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Checkout API
 * POST /api/sessions с телом {} создаёт гостевую сессию. Вставьте data.token в Authorize без слова Bearer. Суммы указаны в копейках. Данные сохраняются после перезапуска. Оплата тестовая.
 * OpenAPI spec version: 1.0.0
 */
import type { GetSandbox200DataCardsItemScenario } from './getSandbox200DataCardsItemScenario';

export type GetSandbox200DataCardsItem = {
  /**
   * @minLength 1
   * @maxLength 500
   */
  id: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  maskedNumber: string;
  scenario: (typeof GetSandbox200DataCardsItemScenario)[keyof typeof GetSandbox200DataCardsItemScenario];
};